Delay event (un)registrations during its triggered event
Store and delay event unregistrations and registrations if the event they are asking about is currently being triggered. This prevent registrations from being skipped or called multiple times for a single event if the registration list changes during the trigger callbacks.

This was easily seen by cancelling an edge creation with some soon-to-be changes.